requirements.txt
streamlit
pandas
numpy
scipy
matplotlib
geopandas
shapely
folium
pykml
lxml
openpyxl
pyproj
geemap
google-auth
llama-parse
llama-index-core
